feat:钩子函数调整
