+++ src/module/mixin/mixin-selection-command.ts
@@ -2,9 +2,9 @@ import * as http from 'http';
 import * as vscode from 'vscode';
 import * as fs from 'fs';
 import * as path from 'path';
-import { withTsType, generateUUID } from '../utils/index';
-import { parseElTableColumnsFromSelection } from '../utils/parse';
-import { properties } from '../global-object/properties';
+import { withTsType, generateUUID } from '../../utils/index';
+import { parseElTableColumnsFromSelection } from '../../utils/parse';
+import { properties } from '../../global-object/properties';
 
 
 // 根据选中内容生成ts类型
